+++ task2/frontend/src/environments/environment.ts
@@ -1,0 +1,5 @@
+export const environment = {
+  production: false,
+  maxTrials: 3,
+  apiHost: 'http://localhost:5043'
+};