+++ changelog.md
@@ -6,3 +6,13 @@
 - task1.csproj
 - readme.md
 - changelog.md
+
+# [0.2.0]
+- added task2 (client side and server side)
+- added session validation and resume functionality
+## Added
+- backend
+- api endpoints(start, validate, submit, cleanup)
+- frontend
+- session validation and resume functionality
+